feat: 로그인/회원가입 관련 추가 수정
