+++ src/api/types/base.ts
@@ -1,0 +1,9 @@
+export interface BaseResponse {
+  isSuccess: boolean;
+  code: string;
+  message: string;
+}
+
+export interface CommonResponse extends BaseResponse {
+  result: string;
+}
