+++ src/api/types/error.ts
@@ -1,0 +1,5 @@
+import { BaseResponse } from './base';
+
+export interface ErrorResponse extends BaseResponse {
+  data?: unknown;
+}
