+++ src/api/types/user.ts
@@ -1,0 +1,18 @@
+import { BaseResponse } from './base';
+
+export interface MyPageResponse extends BaseResponse {
+  result: {
+    name: string;
+    birth: string;
+    gender: 'MALE' | 'FEMALE';
+    phone: string;
+    email: string;
+  };
+}
+
+export interface UpdateMyPageRequest {
+  name: string;
+  birth: string;
+  email: string;
+  password: string;
+}
